# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/{userId}:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get a user's wallet (auto-creates if missing)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/{userId}/ensure:
    post:
      operationId: ensureWallet
      tags: [wallet]
      summary: Ensure a wallet exists, seeding a fixed server-side welcome balance on first create
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/{userId}/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallet]
      summary: List a user's wallet transaction history
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction history (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"
  /wallet/{userId}/recharge:
    post:
      operationId: rechargeWallet
      tags: [wallet]
      summary: Recharge coins by purchasing a coin package
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeInput"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wallet/{userId}/purchase:
    post:
      operationId: purchaseItem
      tags: [wallet]
      summary: Buy a store item, spending from the wallet and granting ownership
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "200":
          description: Result of purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResult"
        "400":
          description: Insufficient balance or already owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wallet/{userId}/claim-task:
    post:
      operationId: claimTask
      tags: [wallet]
      summary: Claim a daily task reward, crediting the wallet
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimTaskInput"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Already claimed today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wallet/{userId}/items:
    get:
      operationId: listUserItems
      tags: [wallet]
      summary: List a user's owned items
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Owned items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserItem"
  /wallet/{userId}/equip:
    post:
      operationId: equipItem
      tags: [wallet]
      summary: Equip an owned item (unequips others of the same type)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipInput"
      responses:
        "200":
          description: Updated owned items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserItem"
        "400":
          description: Item not owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wallet/{userId}/task-claims:
    get:
      operationId: listTaskClaims
      tags: [wallet]
      summary: List the user's task claims for today
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Today's task claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskClaim"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Wallet:
      type: object
      required:
        - userId
        - coins
        - vPoints
      properties:
        userId: { type: string }
        coins: { type: integer }
        vPoints: { type: integer }
    WalletTransaction:
      type: object
      required:
        - id
        - userId
        - currency
        - amount
        - balanceAfter
        - type
        - description
        - refId
        - createdAt
      properties:
        id: { type: integer }
        userId: { type: string }
        currency: { type: string }
        amount: { type: integer }
        balanceAfter: { type: integer }
        type: { type: string }
        description: { type: string }
        refId: { type: string }
        createdAt: { type: string }
    UserItem:
      type: object
      required:
        - id
        - userId
        - itemId
        - equipped
      properties:
        id: { type: integer }
        userId: { type: string }
        itemId: { type: integer }
        equipped: { type: boolean }
    TaskClaim:
      type: object
      required:
        - taskId
        - claimedOn
      properties:
        taskId: { type: integer }
        claimedOn: { type: string }
    PurchaseResult:
      type: object
      required:
        - wallet
        - item
      properties:
        wallet:
          $ref: "#/components/schemas/Wallet"
        item:
          $ref: "#/components/schemas/UserItem"
    RechargeInput:
      type: object
      required:
        - packageId
      properties:
        packageId: { type: integer }
    PurchaseInput:
      type: object
      required:
        - itemId
      properties:
        itemId: { type: integer }
    ClaimTaskInput:
      type: object
      required:
        - taskId
      properties:
        taskId: { type: integer }
    EquipInput:
      type: object
      required:
        - itemId
      properties:
        itemId: { type: integer }
...